pyproject.toml

[project]
name = "conseal"
version = "2024.12"
dependencies = [
    "numpy < 2.0.0",
    "jpeglib >= 1.0.0",
    "scipy",
    "numba",
    "setuptools"
]
requires-python = ">= 3.8"
authors = [
    {name = "Martin Benes", email = "martin.benes@uibk.ac.at"},
    {name = "Benedikt Lorch", email = "benedikt.lorch@uibk.ac.at"}
]
description = "Implementation of modern image steganographic algorithms"
readme = "README.md"
license = {file = "LICENSE"}
keywords = ["steganography", "image", "jpeg"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Developers",
    "Intended Audience :: Education",
    "Intended Audience :: Information Technology",
    "Intended Audience :: Other Audience",
    "License :: OSI Approved :: Mozilla Public License 2.0 (MPL 2.0)",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Topic :: Communications",
    "Topic :: Education",
    "Topic :: Multimedia",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering",
    "Topic :: Scientific/Engineering :: Image Processing",
    "Topic :: Security",
    "Topic :: Utilities"
]

[project.urls]
Homepage = "https://pypi.org/project/conseal/"
Documentation = "https://conseal.readthedocs.io/"
Source = "https://github.com/uibk-uncover/conseal/"

[project.optional-dependencies]
dev = [
    "flake8",
    "parameterized",
    "pillow"
]
deploy = []

[tool.setuptools.packages.find]
include = ["conseal*"]
exclude = ["wheelhouse*"]
namespaces = false

[tool.coverage.report]
fail_under = 80

[tool.coverage.run]
branch = true
include = ["conseal/*"]
command_line = "-m unittest discover -s test/"
